TM-246: рефакторинг сторінки користувача
