+++ client/src/data/mockData.jsx
@@ -1,0 +1,233 @@
+export const mockOffers = [
+  {
+    id: 1,
+    title: 'Валентин Добротворцев',
+    shortDescription:
+      'Організовую подорожі, практикую тілесну терапію, створюю сайти та виготовляю ложки з дерева',
+    description:
+      'Відкритий до безгрошових взаємообмінів через Клуб Організаторів. Практикую бартерний обмін послугами та знаннями.',
+    mainImage: 'https://ko.in.ua/wp-content/uploads/2020/11/logoVD-e1609080023185.jpg',
+    images: [
+      'https://ko.in.ua/wp-content/uploads/2020/11/logoVD-e1609080023185.jpg',
+      'https://ko.in.ua/wp-content/uploads/bfi_thumb/Logo_ST-qiuj3pbukfjvfzmxw5bkil2jcfjecb7f8wphelpuqw.png',
+      'https://ko.in.ua/wp-content/uploads/2020/11/4.0.jpg',
+      'https://ko.in.ua/wp-content/uploads/2020/11/photo_2022-02-10-00.01.49.jpeg',
+    ],
+    address: 'Київ, Україна',
+    contacts: {
+      phone: '+38 (093) 927-72-39',
+      facebook: 'facebook.com/dobrotvorcev',
+      youtube: 'youtube.com/@zkowt',
+      instagram: 'instagram.com/shliah_tvorcia',
+      telegram: 'https://bit.ly/valentin-dobrotvorcev',
+    },
+    sections: {
+      travel: {
+        title: 'Шлях Творця',
+        subtitle: 'Туристичний маршрут від Києва до моря',
+        description:
+          'Започаткував та розвиваю туристичний маршрут через всю Україну. Акцентую увагу на відрізку маршруту від Українки до Канева і далі на захід проробляю нову гілку маршруту.',
+        fullDescription: `Організовую групові та індивідуальні подорожі Шляхом Творця, а також консультую бажаючих пройтися Шляхом самостійно.
+
+В подорожах ми:
+• Пізнаємо рідний край
+• Пізнаємо людей та їх професії
+• Вчимося та товаришуємо
+• Практикуємо духовне просвітлення
+
+**Популярні маршрути:**
+1. Активна прогулянка "Три Купелі" - Київ
+2. Похід "Озеро за лісом" - Київ
+3. Середній похід Шляхом Творця (Українка → Канев)`,
+        projects: [
+          {
+            id: 1,
+            name: 'Шлях Творця 2023',
+            image:
+              'https://ko.in.ua/wp-content/uploads/bfi_thumb/11-qaektgj3y35je8r6in98zkcl70lm26sxsozpwm4trg.jpg',
+            description: 'Середній похід Шляхом Творця, йдемо з Українки до Канева в Пошуках Домів',
+            link: 'https://shliah-tvorcia.ko.in.ua/pohid-shliahom-tvorcia-2023/',
+          },
+        ],
+        videos: [
+          { id: 1, url: 'https://www.youtube.com/embed/bjQ2grA3FIg', title: 'Про Шлях Творця' },
+          { id: 2, url: 'https://www.youtube.com/embed/7XBSp-0EDU8', title: 'Три Купелі' },
+          { id: 3, url: 'https://www.youtube.com/embed/y2RtjdSYX9I', title: 'Озеро за лісом' },
+        ],
+        reports: [
+          {
+            id: 1,
+            title: 'Три Купелі',
+            image:
+              'https://ko.in.ua/wp-content/uploads/bfi_thumb/245271560_4390566621024964_6261619414226046259_n-peeu34k3hefai29isvb4ip2b6nj1nkouzw12c1it9o.jpg',
+            description: 'Активна прогулянка по Києву',
+            date: '10.10.2021',
+            link: 'https://ko.in.ua/tru-kypeli/',
+          },
+          {
+            id: 2,
+            title: 'Озеро за лісом',
+            image:
+              'https://ko.in.ua/wp-content/uploads/bfi_thumb/photo1634493912-1-peqxmb2slm1i03dnea5ga3xwhaua59yaompl8xuegc.jpeg',
+            description: 'Активна прогулянка по Києву',
+            date: '17.10.2021',
+            link: 'https://ko.in.ua/ozero-za-lisom-zvit/',
+          },
+        ],
+        reviews: [
+          {
+            id: 1,
+            user: 'Катерина з Екопростору "Писанка"',
+            avatar:
+              'https://ko.in.ua/wp-content/uploads/2023/08/290686197_1206662836766525_3937860109189581756_n.jpeg',
+            text: 'Валентин зупиняється у нас зі своїми туристами. Він згуртовує навколо себе таких самих добротворчих людей, як і він сам.',
+            date: '2023',
+          },
+          {
+            id: 2,
+            user: 'Annete',
+            avatar:
+              'https://ko.in.ua/wp-content/uploads/2021/10/photo_2021-06-04_01-41-25-150x150.jpg',
+            text: 'Це була чудова прогулянка. Я відкрила для себе нові місця рідного міста. Валентин — чудовий провідник, що розказує цікаву інформацію про відвідані місця.',
+            date: '10.10.2021',
+          },
+        ],
+      },
+      wellness: {
+        title: 'Оздоровлення',
+        subtitle: 'Ретритний Суботник',
+        description:
+          'День внутрішнього очищення через неїдення та взаємодію з Природою через стихії',
+        fullDescription: `Запрошую у гості на програму Суботника, яка може формуватися в залежності від вашого запиту наперед, або у довільній формі по ходу діла. З мене супровід, підтримка, увага, відгук.
+
+**Опції суботника:**
+• Неїдення
+• Інформ детокс
+• Тілесна терапія
+• Мовчання
+• Фізична активність
+• Окунання в холодній воді
+• Взаємодія з вогнем
+• Прогулянка лісом
+• Їжа з печі`,
+        videos: [
+          { id: 1, url: 'https://www.youtube.com/embed/b4nnOEya3HA', title: 'Суботник' },
+          { id: 2, url: 'https://www.youtube.com/embed/BbnBus4Pxts', title: 'Ретрит' },
+        ],
+        reviews: [],
+      },
+      websites: {
+        title: 'Сайти',
+        subtitle: 'Створення сайтів на WordPress',
+        description:
+          'Створюю сайти під ключ на wordpress: сайти-візитки та лендінги. По дизайну сайта — консультую!',
+        fullDescription: `**Послуги:**
+• Сайти-візитки
+• Лендінги
+• Консультації по дизайну
+• Розробка під ключ
+
+**Портфоліо:**
+1. Authentic - http://authentic.com.ua/
+2. Tesla Garage - https://teslagarage.com.ua/
+3. Шлях Творця - https://shliah-tvorcia.com.ua/
+4. Клуб Організаторів - https://ko.in.ua/`,
+        portfolio: [
+          { name: 'Authentic', url: 'http://authentic.com.ua/' },
+          { name: 'Tesla Garage', url: 'https://teslagarage.com.ua/' },
+          { name: 'Шлях Творця', url: 'https://shliah-tvorcia.com.ua/' },
+          { name: 'Клуб Організаторів', url: 'https://ko.in.ua/' },
+        ],
+        reviews: [],
+      },
+      spoons: {
+        title: 'Ложки з дерева',
+        subtitle: "Виготовлення дерев'яних ложок на замовлення",
+        description:
+          'Обожнюю робити ложки з дерева для людей на замовлення. Під час роботи думаю про людину і ложка виходить особливою та суто індивідуальною.',
+        fullDescription: `**Переваги дерев'яних ложок:**
+• Не окислюються (на відміну від металу)
+• Не збивають емаль із зубів
+• Натуральний матеріал
+• Енергетично наповнені
+
+**Цінність:** Від 500 грн`,
+        images: [
+          'https://ko.in.ua/wp-content/uploads/bfi_thumb/photo_2022-02-10-00.01.44-pka0iek2egucce3w38dlkc4hrpyt6js7rhe6rwut9o.jpeg',
+          'https://ko.in.ua/wp-content/uploads/bfi_thumb/photo_2022-02-10-00.01.46-pka0igfqs4wwzm15s96upbneyhpjlxzofqp5qgs0x8.jpeg',
+          'https://ko.in.ua/wp-content/uploads/bfi_thumb/photo_2022-02-10-00.01.48-pka0igfqs4wwzm15s96upbneyhpjlxzofqp5qgs0x8.jpeg',
+        ],
+        reviews: [],
+      },
+    },
+    category: 'Клуб Організаторів',
+    author: {
+      name: 'Валентин Добротворцев',
+      avatar: 'https://ko.in.ua/wp-content/uploads/2020/11/logoVD-e1609080023185.jpg',
+      rating: 4.8,
+      reviews: 24,
+      experience: 'Організатор подорожей, тілесний терапевт, веб-розробник',
+      location: 'Київ, Україна',
+      memberSince: '2020 року',
+      contacts: {
+        phone: '+38 (093) 927-72-39',
+        telegram: 'https://bit.ly/valentin-dobrotvorcev',
+        facebook: 'facebook.com/dobrotvorcev',
+        youtube: 'youtube.com/@zkowt',
+      },
+    },
+    price: 'Бартер / Вільний внесок',
+    date: 'Оновлено 07.02.2024',
+    tags: ['Подорожі', 'Тілесна терапія', 'Веб-розробка', 'Деревообробка', 'Бартер'],
+    details: {
+      format: 'Онлайн/Офлайн',
+      exchange: 'Бартерний обмін',
+      location: 'Київ та Україна',
+      languages: 'Українська, Російська',
+    },
+  },
+];
+
+// Мокові дані для маркерів (для мапи)
+export const mockMarkers = [
+  {
+    id: 1,
+    lat: 50.4501,
+    lng: 30.5234,
+    title: 'Київ',
+    description: 'Столиця України',
+    type: 'capital',
+  },
+  {
+    id: 2,
+    lat: 49.8397,
+    lng: 24.0297,
+    title: 'Львів',
+    description: 'Місто Лева',
+    type: 'city',
+  },
+  {
+    id: 3,
+    lat: 48.9226,
+    lng: 24.7111,
+    title: 'Івано-Франківськ',
+    description: 'Ворота Карпат',
+    type: 'city',
+  },
+  {
+    id: 4,
+    lat: 48.3794,
+    lng: 31.1656,
+    title: 'Українка',
+    description: 'Початок маршруту "Шлях Творця"',
+    type: 'route_start',
+  },
+  {
+    id: 5,
+    lat: 49.75,
+    lng: 31.45,
+    title: 'Канів',
+    description: "Кінець маршруту 'Шлях Творця'",
+    type: 'route_end',
+  },
+];
