+++ client/src/modules/announcements/announcementsDetail/LoadingSpinner.jsx
@@ -1,0 +1,10 @@
+export const LoadingSpinner = () => {
+  return (
+    <div className="min-h-screen bg-gray-50 flex items-center justify-center">
+      <div className="text-center">
+        <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-[#e65000] mx-auto"></div>
+        <p className="mt-4 text-gray-600">Завантаження...</p>
+      </div>
+    </div>
+  );
+};
